feat(site.yml): add playbook to deploy and configure Azure environment
This commit adds a new playbook called "site.yml" which is responsible for deploying and configuring the Azure environment. The playbook consists of two roles: "azure-rg" and "azure-virtualnetwork".

The "azure-rg" role is responsible for creating the resource group in Azure. It uses the "azure_rm_resourcegroup" module to create the resource group with the specified name and location. The role also registers the output of the resource group creation for debugging purposes.

The "azure-virtualnetwork" role is responsible for creating the virtual network in Azure. It uses the "azure_rm_virtualnetwork" module to create the virtual network with the specified name and address prefixes. The role also registers the output of the virtual network creation for debugging purposes.

Both roles use variables defined in the "group_vars/common.yml" file to configure the resource group and virtual network.

Note: Uncomment the "tasks" section in the playbook to enable debugging messages.