feat(site.yml): add playbook to deploy and configure the Azure Environment
This commit adds a new playbook called "site.yml" which is responsible for deploying and configuring the Azure Environment. The playbook is executed on the localhost and gathers facts about the system.

The playbook includes two roles: "azure-rg" and "azure-virtualnetwork". The "azure-rg" role is responsible for creating the resource group, while the "azure-virtualnetwork" role is responsible for creating the virtual network.

The playbook also includes a commented-out task that prints debug information about the system.

This commit is part of the process of setting up the basic infrastructure for the application.